requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.8
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
requests>=2.32.0
anyio>=4.0.0
alembic>=1.13.0
